+++ esp_cam_sensor/sensors/mira220/private_include/mira220_settings.h
@@ -1,0 +1,23 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include <stdio.h>
+#include <stdint.h>
+#include <sdkconfig.h>
+#include "mira220_regs.h"
+#include "mira220_types.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "mira220_mipi_2lane_24Minput_1024x600_raw8_15fps.h"
+
+#ifdef __cplusplus
+}
+#endif